LayoutTests/fast/css/parsing-css-not-9-expected.txt

Test the parsing of :not(selectorList) for querySelector and style. This test focus on pseudo elements nested inside other functional pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


3 :not(valid, valid, invalid) selectors
selector ':not(foobar,foobar,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,foobar,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,foobar,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,foobar,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,foobar,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,foobar,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,foobar,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,foobar,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,foobar,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,foobar,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,foobar,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,foobar,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,foobar,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,foobar,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,.class,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,.class,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,.class,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,.class,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,.class,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,.class,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,.class,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,.class,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,.class,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,.class,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,.class,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,.class,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,.class,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,.class,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:last-child,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:last-child,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:last-child,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:last-child,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:last-child,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:last-child,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:last-child,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:last-child,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:last-child,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:last-child,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:last-child,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:last-child,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:last-child,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:last-child,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,foobar#id.class,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,#id:empty,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,#id:empty,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,#id:empty,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,#id:empty,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,#id:empty,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,#id:empty,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,#id:empty,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,#id:empty,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,#id:empty,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,#id:empty,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,a b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,a b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,a b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,a b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,a b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,a b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,a b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,a b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,a b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,a b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,a b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,a b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,a b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,a b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,a ~ b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,a ~ b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,a ~ b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,a ~ b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,a ~ b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,a ~ b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,a ~ b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,a ~ b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,a ~ b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,a ~ b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar,:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,.class,:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,.class,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,.class,:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,.class,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,.class,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,.class,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,.class,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,.class,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,.class,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,.class,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,.class,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,.class,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,.class,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,.class,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:last-child,:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:last-child,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:last-child,:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:last-child,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:last-child,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:last-child,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:last-child,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,:last-child,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,:last-child,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:last-child,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:last-child,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:last-child,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:last-child,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:last-child,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,foobar#id.class,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,foobar#id.class,:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,foobar#id.class,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,foobar#id.class,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,foobar#id.class,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,#id:empty,:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,#id:empty,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,#id:empty,:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,#id:empty,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,#id:empty,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,#id:empty,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,#id:empty,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,#id:empty,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,#id:empty,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,#id:empty,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,a b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,a b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,a b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,a b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,a b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,a b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,a b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,a b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,a b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,a b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,a b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,a b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,a b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,a b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,a ~ b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,a ~ b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,a ~ b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,a ~ b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,a ~ b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,a ~ b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,a ~ b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,a ~ b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,a ~ b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,a ~ b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(.class,:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:last-child,:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:last-child,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:last-child,:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:last-child,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:last-child,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:last-child,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:last-child,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:last-child,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:last-child,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:last-child,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:last-child,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:last-child,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:last-child,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:last-child,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,foobar#id.class,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,#id:empty,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,#id:empty,:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,#id:empty,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,#id:empty,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,a b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,a b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,a b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,a b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,a b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,a b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,a b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,a b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,a b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,a b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,a b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,a b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,a b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,a b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,a ~ b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,a ~ b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,a ~ b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,a ~ b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:last-child,:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,foobar#id.class,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,#id:empty,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,a b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,a b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,a b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,a b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,a b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,a ~ b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(foobar#id.class,:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:is(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:is(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,#id:empty,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,a b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,a b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,a b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,a b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,a b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,a b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,a b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,a b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,a b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,a b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,a b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,a b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,a b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,a b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,a ~ b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(#id:empty,:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,a b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(a b,a b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a b,a b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(a b,a b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a b,a b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a b,a b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a b,a b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a b,a b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a b,a b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a b,a b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a b,a b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a b,a b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a b,a b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,a b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,a ~ b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(a b,a ~ b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a b,a ~ b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(a b,a ~ b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a b,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a b,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a b,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a b,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a b,a ~ b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a b,a ~ b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a b,a ~ b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a b,a ~ b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a b,a ~ b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,a ~ b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c::after)) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c:after)) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a b,:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:is(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:is(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,a ~ b,:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(a ~ b,:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:-webkit-any(.selector,#tama,#hanayo,#midoriko),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:-webkit-any(.selector,#tama,#hanayo,#midoriko),:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:nth-child(even of a, b, c),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:nth-child(even of a, b, c),:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:is(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:not(:is(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:is(a, b, c::after)) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:matches(a, b, c::after)) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:is(a, b, c:after)) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:matches(a, b, c:after)) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, ::before))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, ::before))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :is(f, g, :before))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:nth-child(2n+1 of d, e, :matches(f, g, :before))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c::after))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c::after))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:not(:is(a, b, c:after))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:not(:matches(a, b, c:after))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:is(a, b, :is(c, d, :is(e, f, :is(g, h, i::after))))) { }' parsed: YES
selector ':not(:not(:matches(cocoa, cappuccino)),:not(:matches(cocoa, cappuccino)),:matches(a, b, :matches(c, d, :matches(e, f, :matches(g, h, i::after))))) { }' parsed: YES
PASS successfullyParsed is true

TEST COMPLETE
